+++ android-record/pom.xml
@@ -1,96 +1,97 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <!-- This module was also published with a richer model, Gradle metadata,  -->
-    <!-- which should be used instead. Do not delete the following line which  -->
-    <!-- is to indicate to Gradle or any Gradle module metadata file consumer  -->
-    <!-- that they should prefer consuming it instead. -->
-    <!-- do_not_remove: published-with-gradle-metadata -->
-    <parent>
-        <groupId>com.fasterxml.jackson.module</groupId>
-        <artifactId>jackson-modules-base</artifactId>
-        <version>2.16.0-SNAPSHOT</version>
-    </parent>
-    <modelVersion>4.0.0</modelVersion>
+  <!-- This module was also published with a richer model, Gradle metadata,  -->
+  <!-- which should be used instead. Do not delete the following line which  -->
+  <!-- is to indicate to Gradle or any Gradle module metadata file consumer  -->
+  <!-- that they should prefer consuming it instead. -->
+  <!-- do_not_remove: published-with-gradle-metadata -->
+  <parent>
+    <groupId>com.fasterxml.jackson.module</groupId>
+    <artifactId>jackson-modules-base</artifactId>
+    <version>2.16.0-SNAPSHOT</version>
+  </parent>
+  <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>jackson-module-android-record</artifactId>
-    <name>Jackson module: Android Record Support</name>
-    <packaging>bundle</packaging>
+  <artifactId>jackson-module-android-record</artifactId>
+  <name>Jackson module: Android Record Support</name>
+  <packaging>bundle</packaging>
 
-    <description>Support deserialization into records on Android</description>
-    <url>https://github.com/FasterXML/jackson-modules-base</url>
+  <description>Support deserialization into records on Android</description>
+  <url>https://github.com/FasterXML/jackson-modules-base</url>
 
-    <licenses>
-        <license>
-            <name>The Apache Software License, Version 2.0</name>
-            <url>https://www.apache.org/licenses/LICENSE-2.0.txt</url>
-            <distribution>repo</distribution>
-        </license>
-    </licenses>
+  <licenses>
+    <license>
+      <name>The Apache Software License, Version 2.0</name>
+      <url>https://www.apache.org/licenses/LICENSE-2.0.txt</url>
+      <distribution>repo</distribution>
+    </license>
+  </licenses>
 
-    <properties>
-        <!-- Generate PackageVersion.java into this directory. -->
-        <packageVersion.dir>com/fasterxml/jackson/module/androidrecord</packageVersion.dir>
-        <packageVersion.package>com.fasterxml.jackson.module.androidrecord</packageVersion.package>
-    </properties>
+  <properties>
+    <!-- Generate PackageVersion.java into this directory. -->
+    <packageVersion.dir>com/fasterxml/jackson/module/androidrecord</packageVersion.dir>
+    <packageVersion.package>com.fasterxml.jackson.module.androidrecord</packageVersion.package>
+  </properties>
 
-    <dependencies>
-        <dependency>
-            <groupId>com.fasterxml.jackson.core</groupId>
-            <artifactId>jackson-core</artifactId>
-        </dependency>
-      <dependency>
-          <groupId>com.fasterxml.jackson.core</groupId>
-          <artifactId>jackson-annotations</artifactId>
-      </dependency>
-        <dependency>
-            <groupId>com.fasterxml.jackson.core</groupId>
-            <artifactId>jackson-databind</artifactId>
-        </dependency>
-    </dependencies>
+  <dependencies>
+    <dependency>
+      <groupId>com.fasterxml.jackson.core</groupId>
+      <artifactId>jackson-core</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>com.fasterxml.jackson.core</groupId>
+      <artifactId>jackson-annotations</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>com.fasterxml.jackson.core</groupId>
+      <artifactId>jackson-databind</artifactId>
+    </dependency>
+  </dependencies>
 
-    <build>
-        <plugins>
-          <plugin>
-               <groupId>org.apache.maven.plugins</groupId>
-               <artifactId>maven-compiler-plugin</artifactId>
-               <inherited>true</inherited>
-               <configuration>
-                 <compilerArgs>
-                   <arg>-parameters</arg>
-                 </compilerArgs>
-                 <fork>true</fork>
-                 <useIncrementalCompilation>true</useIncrementalCompilation>
-               </configuration>
-           </plugin>
+  <build>
+    <plugins>
+      <plugin>
+        <groupId>org.apache.maven.plugins</groupId>
+        <artifactId>maven-compiler-plugin</artifactId>
+        <inherited>true</inherited>
+        <configuration>
+          <compilerArgs>
+            <!-- Module uses constructor parameter names (and types) to identify the canonical constructor -->
+            <arg>-parameters</arg>
+          </compilerArgs>
+          <fork>true</fork>
+          <useIncrementalCompilation>true</useIncrementalCompilation>
+        </configuration>
+      </plugin>
 
-            <plugin>
-                <!-- Inherited from oss-base. Generate PackageVersion.java.-->
-                <groupId>com.google.code.maven-replacer-plugin</groupId>
-                <artifactId>replacer</artifactId>
-                <executions>
-                    <execution>
-                        <id>process-packageVersion</id>
-                        <phase>generate-sources</phase>
-                    </execution>
-                </executions>
-            </plugin>
+      <plugin>
+        <!-- Inherited from oss-base. Generate PackageVersion.java.-->
+        <groupId>com.google.code.maven-replacer-plugin</groupId>
+        <artifactId>replacer</artifactId>
+        <executions>
+          <execution>
+            <id>process-packageVersion</id>
+            <phase>generate-sources</phase>
+          </execution>
+        </executions>
+      </plugin>
 
-            <plugin>
-                <groupId>org.moditect</groupId>
-                <artifactId>moditect-maven-plugin</artifactId>
-            </plugin>
+      <plugin>
+        <groupId>org.moditect</groupId>
+        <artifactId>moditect-maven-plugin</artifactId>
+      </plugin>
 
-            <plugin>
-              <groupId>org.codehaus.mojo</groupId>
-              <artifactId>animal-sniffer-maven-plugin</artifactId>
-              <version>${version.plugin.animal-sniffer}</version>
-              <configuration>
-                <signature>
-                  <groupId>com.toasttab.android</groupId>
-                  <artifactId>gummy-bears-api-${version.android.sdk}</artifactId>
-                  <version>${version.android.sdk.signature}</version>
-                </signature>
-              </configuration>
-            </plugin>
-        </plugins>
-    </build>
+      <plugin>
+        <groupId>org.codehaus.mojo</groupId>
+        <artifactId>animal-sniffer-maven-plugin</artifactId>
+        <version>${version.plugin.animal-sniffer}</version>
+        <configuration>
+          <signature>
+            <groupId>com.toasttab.android</groupId>
+            <artifactId>gummy-bears-api-${version.android.sdk}</artifactId>
+            <version>${version.android.sdk.signature}</version>
+          </signature>
+        </configuration>
+      </plugin>
+    </plugins>
+  </build>
 </project>
