+++ android-record/src/test/java/com/fasterxml/jackson/module/androidrecord/RecordBasicsTest.java
@@ -4,16 +4,12 @@ import com.android.tools.r8.RecordTag;
 import com.fasterxml.jackson.annotation.JacksonInject;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.InjectableValues;
-import com.fasterxml.jackson.databind.JavaType;
 import com.fasterxml.jackson.databind.MapperFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.SerializationFeature;
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 import com.fasterxml.jackson.databind.json.JsonMapper;
-import com.fasterxml.jackson.databind.type.TypeFactory;
-import com.fasterxml.jackson.databind.util.Converter;
 
 import java.util.Collections;
 import java.util.LinkedHashMap;
@@ -104,26 +100,6 @@ public class RecordBasicsTest extends BaseMapTest
     }
   }
 
-  static final class RecordWithHeaderInject extends RecordTag {
-    private final int id;
-    @JacksonInject
-    private final String name;
-
-    RecordWithHeaderInject(int id, @JacksonInject String name) {
-      this.id = id;
-      this.name = name;
-    }
-
-    public int id() {
-      return id;
-    }
-
-    @JacksonInject
-    public String name() {
-      return name;
-    }
-  }
-
   static final class RecordWithConstructorInject extends RecordTag {
     private final int id;
     private final String name;
@@ -179,46 +155,6 @@ public class RecordBasicsTest extends BaseMapTest
       }
       SnakeRecord that = (SnakeRecord) o;
       return Objects.equals(myId, that.myId) && Objects.equals(myValue, that.myValue);
-    }
-  }
-
-  static final class RecordWithJsonDeserialize extends RecordTag {
-    private final int id;
-    @JsonDeserialize(converter = StringTrimmer.class)
-    private final String name;
-
-    RecordWithJsonDeserialize(int id, @JsonDeserialize(converter = StringTrimmer.class) String name) {
-      this.id = id;
-      this.name = name;
-    }
-
-    public int id() {
-      return id;
-    }
-
-    @JsonDeserialize(converter = StringTrimmer.class)
-    public String name() {
-      return name;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-      if (this == o) {
-        return true;
-      }
-      if (!(o instanceof RecordWithJsonDeserialize)) {
-        return false;
-      }
-      RecordWithJsonDeserialize that = (RecordWithJsonDeserialize) o;
-      return id == that.id && Objects.equals(name, that.name);
-    }
-
-    @Override
-    public String toString() {
-      return "RecordWithJsonDeserialize{" +
-              "id=" + id +
-              ", name='" + name + '\'' +
-              '}';
     }
   }
 
@@ -433,35 +369,6 @@ public class RecordBasicsTest extends BaseMapTest
         assertEquals(EXP, MAPPER.readValue(json, Object.class));
     }
 
-    // Fails in BasicDeserializerFactory._addImplicitConstructorCreators before reaching module logic:
-    // Invalid type definition for type `com.fasterxml.jackson.module.androidrecord.RecordBasicsTest$RecordWithRename`: Argument #0 of constructor [constructor for `com.fasterxml.jackson.module.androidrecord.RecordBasicsTest$RecordWithRename` (2 args), annotations: [null] has no property name annotation; must have name when multiple-parameter constructor annotated as Creator
-    public void notestDeserializeJsonRename() throws Exception {
-        RecordWithRename value = MAPPER.readValue("{\"id\":123,\"rename\":\"Bob\"}",
-                RecordWithRename.class);
-        assertEquals(new RecordWithRename(123, "Bob"), value);
-    }
-
-    // Fails by deserializing successfully, even though annotations on header are "propagated" to the field
-    /**
-     * This test-case is just for documentation purpose:
-     * GOTCHA: Annotations on header will be propagated to the field, leading to this failure.
-     *
-     * @see #testDeserializeConstructorInjectRecord()
-     */
-    public void notestDeserializeHeaderInjectRecord_WillFail() throws Exception {
-        MAPPER.setInjectableValues(new InjectableValues.Std().addValue(String.class, "Bob"));
-
-        try {
-            MAPPER.readValue("{\"id\":123}", RecordWithHeaderInject.class);
-
-            fail("should not pass");
-        } catch (IllegalArgumentException e) {
-            verifyException(e, "RecordWithHeaderInject#name");
-            verifyException(e, "Can not set final java.lang.String field");
-        }
-    }
-
-
     public void testDeserializeConstructorInjectRecord() throws Exception {
         MAPPER.setInjectableValues(new InjectableValues.Std().addValue(String.class, "Bob"));
 
@@ -485,19 +392,6 @@ public class RecordBasicsTest extends BaseMapTest
 
         SnakeRecord output = MAPPER.readValue(json, SnakeRecord.class);
         assertEquals(input, output);
-    }
-
-    /*
-    /**********************************************************************
-    /* Test methods, JsonDeserialize
-    /**********************************************************************
-     */
-
-    // Fails: converter not applied
-    public void notestDeserializeJsonDeserializeRecord() throws Exception {
-        RecordWithJsonDeserialize value = MAPPER.readValue("{\"id\":123,\"name\":\"   Bob   \"}", RecordWithJsonDeserialize.class);
-
-        assertEquals(new RecordWithJsonDeserialize(123, "Bob"), value);
     }
 
     /*
@@ -553,29 +447,11 @@ public class RecordBasicsTest extends BaseMapTest
     /**********************************************************************
      */
 
-    private Map<String,Object> map(String key1, Object value1,
-            String key2, Object value2) {
+    private static Map<String,Object> map(String key1, Object value1,
+                                          String key2, Object value2) {
         final Map<String, Object> result = new LinkedHashMap<>();
         result.put(key1, value1);
         result.put(key2, value2);
         return result;
-    }
-
-    public static class StringTrimmer implements Converter<String, String> {
-
-        @Override
-        public String convert(String value) {
-            return value.trim();
-        }
-
-        @Override
-        public JavaType getInputType(TypeFactory typeFactory) {
-            return typeFactory.constructType(String.class);
-        }
-
-        @Override
-        public JavaType getOutputType(TypeFactory typeFactory) {
-            return typeFactory.constructType(String.class);
-        }
     }
 }
